backend/pyproject.toml
[tool.poetry]
name = "hackaton"
version = "0.1.0"
description = ""
authors = ["y.romanchuk <y.romanchuk@bi.zone>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.37"
alembic = "^1.8.0"
psycopg2 = "^2.9.3"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.5"
fastapi-pagination = { extras = ["sqlalchemy"], version = "^0.9.3" }
pydantic = {extras = ["email"], version = "^1.9.1"}
uvicorn = {extras = ["standard"], version = "^0.17.6"}
gunicorn = "^20.1.0"
ipython = "^8.4.0"
pandas = "^1.4.2"

[tool.poetry.dev-dependencies]
sqlalchemy2-stubs = "^0.0.2-alpha.23"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py38"]
exclude = '''
/(
    \.git
  | \.venv(.*)
  | venv
  | migrations
)/
'''

[tool.isort]
line_length = 120
skip_glob = [".venv*", "venv*", "migrations"]
# add for black compatible, more info: https://black.readthedocs.io/en/stable/the_black_code_style.html
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true